ResourceGroupVisualMap
This PowerShell script uses the Azure Resource Graph API to query an Azure resource group and generate a visual map of its resources. The script displays the resources and their dependencies in a graphical format, making it easier to understand how the resources are related and how they interact with each other.

To use this script, you will need to have Azure PowerShell and the Azure Resource Graph module installed on your system. You will also need to have permissions to access the resource group that you want to visualize.  You will need to change Demo-1, Demo-2, Demo-3 to resource groups to get it to work.